major improvements in C API: use caml value pointers instead of values
This free the need of C programmer of having to deal with caml values, and in
particular of having to worry about local/global roots to avoid irritating the
caml GC. Now all value protection happens *inside* the bindings, transparent
for the programmer.

Still TODO: remove the need of invoking caml_startup(), so that caml headers
will no longer be needed to build CUDF-C programs.