Alter DESTDIR for Cygwin building
If DESTDIR isn't defined, various directories in the Makefile end up
prefixed //. While this shouldn't be a problem on most Unices, POSIX
does allow // to have special meaning and Cygwin uses it represent the
network environment (i.e. // is not the same as /).

Alter DESTDIR so that the trailing slash is added only if DESTDIR has
been defined.

Signed-off-by: Stefano Zacchiroli <zack@upsilon.cc>